Move TestTime to own component
