+++ client/src/components/TestValuePicker.tsx
@@ -1,31 +1,16 @@
-import React, { ReactElement, useEffect } from "react";
+import React, { ReactElement } from "react";
 import styled from "styled-components";
 import { t } from "i18next";
-import _ from "lodash";
 import { TestTimeSelector } from "client/test/test-components/TestTimeSelector";
 import { SignupStrategySelector } from "client/test/test-components/SignupStrategySelector";
 import { Accordion } from "client/components/Accordion";
-import { testTimes } from "client/test/test-components/testComponentUtils";
-import { useAppDispatch, useAppSelector } from "client/utils/hooks";
-import { submitSetTestSettings } from "client/test/test-settings/testSettingsThunks";
+import { useAppSelector } from "client/utils/hooks";
 import { config } from "client/config";
 
 export const TestValuePicker = (): ReactElement => {
-  const dispatch = useAppDispatch();
-
   const testTime: string = useAppSelector(
     (state) => state.testSettings.testTime
   );
-
-  useEffect(() => {
-    const setInitialTestTime = async (): Promise<void> => {
-      const defaultTestTime = _.first(testTimes);
-      if (!testTime && defaultTestTime) {
-        await dispatch(submitSetTestSettings({ testTime: defaultTestTime }));
-      }
-    };
-    setInitialTestTime();
-  });
 
   return (
     <div>
